fix(extension): resolved sonarcloud issues
